+++ microservicio/dominio/src/main/java/com/ceiba/pedido/modelo/dto/DtoPedido.java
@@ -17,6 +17,6 @@ public class DtoPedido {
     private String destino;
     private Integer flete;
     private Integer precio;
-    private double descuento;
+    private Integer descuento;
     private double total;
 }
